move the encoded package down to ledger/encoded
