+++ ledger/store/encodedAccountsIter.go
@@ -21,7 +21,7 @@ import (
 	"database/sql"
 
 	"github.com/algorand/go-algorand/data/basics"
-	"github.com/algorand/go-algorand/encoded"
+	"github.com/algorand/go-algorand/ledger/encoded"
 	"github.com/algorand/msgp/msgp"
 )
